Resolves failing Time Tracking spec.
The Harvest API seems to require that a deactivated task be activated
before it is associated with a project (given the task should be active
at a point in the future).

I was only able to activate a deactivated task via the Harvest web
interface when I assigned it to a project while it was deactivated.