Adds tests for Invoice due-at field.
The due at column is directly related to the due-at-human-format
column. Presently, if a custom due-at date is desired, the
due-at-human-format column must be "custom"

The Harvest support team indicated that they may change this behavior
to automatically change the due-at-human-format field to custom
whenever a due-at date is specified.

Additionally, the date format should be the same as the date format
specified on the account.